week4/requirements.txt
langgraph>=0.0.30
langchain-openai>=0.1.0
pydantic
fastapi
uvicorn
python-dotenv
